fix(visualization): Support multiple analysis_geometry and user_data
This commit also changed matching_method to be a property that is constant across all data sets on an AnalysisGeometry object (instead of a method that can be different for different data sets).

We're also removing the ability to color Polyface faces or vertices with values, asking for users to convert these to a mesh first and thereby simplifying our translation methods.